Fixed UnicodeSequence
