+++ unicode/UnicodeSequence.cs
@@ -8,11 +8,15 @@ namespace NeoSmart.Unicode
     /// <summary>
     /// A UnicodeSequence is a combination of one or more codepoints.
     /// </summary>
-    public struct UnicodeSequence : IComparable<UnicodeSequence>, IEquatable<UnicodeSequence>, IEquatable<string>
+    public class UnicodeSequence : IComparable<UnicodeSequence>, IEquatable<UnicodeSequence>, IEquatable<string>, IEqualityComparer<UnicodeSequence>
     {
-        static readonly Codepoint[] NoCodepoints = new Codepoint[] { };
         private readonly Codepoint[] _codepoints;
-        public IEnumerable<Codepoint> Codepoints => _codepoints ?? NoCodepoints;
+        public IEnumerable<Codepoint> Codepoints => _codepoints;
+
+        private UnicodeSequence()
+        {
+
+        }
 
         public UnicodeSequence(string sequence)
         {
@@ -44,11 +48,6 @@ namespace NeoSmart.Unicode
                 var values = sequence.Split(new[] { ',', ' ' }, StringSplitOptions.RemoveEmptyEntries);
                 _codepoints = values.Select(x => new Codepoint(x)).ToArray();
             }
-        }
-
-        public UnicodeSequence(params Codepoint[] codepoints)
-        {
-            _codepoints = codepoints;
         }
 
         public UnicodeSequence(IEnumerable<Codepoint> codepoints)
@@ -117,21 +116,6 @@ namespace NeoSmart.Unicode
 
         public int CompareTo(UnicodeSequence other)
         {
-            if (_codepoints is null && other._codepoints is null)
-            {
-                return 0;
-            }
-
-            if (_codepoints is null)
-            {
-                return -1;
-            }
-
-            if (other._codepoints is null)
-            {
-                return 1;
-            }
-
             for (int i = 0; i < Math.Min(_codepoints.Length, other._codepoints.Length); ++i)
             {
                 if (_codepoints[i] != other._codepoints[i])
@@ -176,7 +160,7 @@ namespace NeoSmart.Unicode
 
         public static bool operator ==(UnicodeSequence a, UnicodeSequence b)
         {
-            return a.Equals(b);
+            return (a is null && b is null) || (!(a is null || b is null) && a.Equals(b));
         }
 
         public static bool operator !=(UnicodeSequence a, UnicodeSequence b)
@@ -188,7 +172,7 @@ namespace NeoSmart.Unicode
         {
             if (b is UnicodeSequence)
             {
-                return Equals((UnicodeSequence)b);
+                return Equals(b as UnicodeSequence);
             }
             return base.Equals(b);
         }
@@ -206,6 +190,11 @@ namespace NeoSmart.Unicode
         public bool Equals(string other)
         {
             return !(other is null) && other.Codepoints().SequenceEqual(_codepoints);
+        }
+
+        public int GetHashCode(UnicodeSequence obj)
+        {
+            return obj.GetHashCode();
         }
     }
 }
